transfered gameloop logic into useGameLoop.js
