+++ src/hooks/useCreateLobby.js
@@ -8,7 +8,7 @@ const useCreateLobby = (socket) => {
   const [lobbyInfo, setLobbyInfo] = useState(null);
   const [players, setPlayers] = useState([]);
   const [userProfile, setUserProfile] = useState({});
-  const [gameState, setGameState] = useState('LOBBY'); 
+  // const [gameState, setGameState] = useState('LOBBY'); 
 
   useEffect(() => {
     // Listen for new host created:
@@ -43,25 +43,9 @@ const useCreateLobby = (socket) => {
       setPlayers(playersData);
     });
 
-    socket.on('startGame', () => {
-      // set state for the gameloop! we playin now bishes
-      setGameState('GAMELOOP'); 
-    });
-
-    socket.on('finalScore', () => {
-      // set state for the gameloop! we playin now bishes
-      setGameState('FINALSCORE'); 
-    });
-
   }, [socket]);
 
-  const startGame = function () {
-    // tell the server that the host has started the game
-    socket.emit('startGame', userProfile.code);
-
-  }
-
-  return {lobbyInfo, players, userProfile, startGame, gameState} ;
+  return {lobbyInfo, players, userProfile} ;
 
 };
